requirements.txt
rich
speedtest-cli
psutil
